# Wrap simulation results writing in TransactionContext